refactor(image-sequence): move helpers to libs
